requirements.txt
asgiref==3.7.2
Babel==2.13.0
certifi==2023.7.22
charset-normalizer==3.3.0
defusedxml==0.7.1
diff-match-patch==20230430
Django==4.2.4
django-appconf==1.0.5
django-ckeditor==6.7.0
django-cleanup==8.0.0
django-cors-headers==4.2.0
django-countries==7.5.1
django-facebook==6.0.3
django-filter==23.3
django-haystack==3.2.1
django-imagekit==5.0.0
django-import-export==3.3.1
django-js-asset==2.1.0
django-measurement==3.2.4
django-model-utils==4.3.1
django-modeltranslation==0.18.11
django-mptt==0.14.0
django-nested-admin==4.0.2
django-parler==2.3
django-phonenumber-field==7.2.0
django-phonenumbers==1.0.1
django-polymorphic==3.1.0
django-polymorphic-tree==2.1
django-prices==2.3.0
django-rosetta==0.9.9
django-tag-parser==3.2
django-widget-tweaks==1.5.0
djangorestframework==3.14.0
enmerkar==0.7.1
et-xmlfile==1.1.0
facebook-sdk==3.1.0
ffmpeg-python==0.2.0
future==0.18.3
gunicorn==21.2.0
idna==3.4
MarkupPy==1.14
measurement==3.2.2
mpmath==1.3.0
odfpy==1.4.1
openpyxl==3.1.2
packaging==23.1
phonenumbers==8.13.22
pilkit==3.0
Pillow==10.0.1
polib==1.2.0
prices==1.1.1
python-monkey-business==1.0.0
pytz==2023.3
PyYAML==6.0.1
requests==2.31.0
six==1.16.0
sqlparse==0.4.4
sympy==1.12
tablib==3.5.0
typing_extensions==4.8.0
tzdata==2023.3
Unidecode==1.3.7
urllib3==2.0.6
whitenoise==6.5.0
Whoosh==2.7.4
xlrd==2.0.1
xlwt==1.3.0